changed user entity

+++ roadmap/4-web-frameworks/1-basic/expense-tracker-api/src/main/java/org/project/expensetrackerapi/model/Expense.java
@@ -29,7 +29,7 @@ public class Expense {
     private Category category;
 
     @Column(name = "cost")
-    private int cost;
+    private double cost;
 
     @Column(name = "date")
     private LocalDate date;
